refactor: change from camel case to snake case
