+++ deploy/oracles/aggregators/lp_chainlink_aggregator.ts
@@ -1,8 +1,9 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployFunction } from "hardhat-deploy/types";
 import { ethers, upgrades } from "hardhat";
-import { OffChainOracle, OffChainOracle__factory } from "../../typechain/v8";
-import { withNetworkFile } from "../../utils";
+import { LPChainlinkAggregator, LPChainlinkAggregator__factory } from "../../../typechain/v8";
+import { getConfig, withNetworkFile } from "../../../utils";
+import { LatteSwapPair__factory } from "@latteswap/latteswap-contract/compiled-typechain";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   /*
@@ -15,24 +16,28 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   Check all variables below before execute the deployment script
   */
   const deployer = (await ethers.getSigners())[0];
-  const FEEDER = "0x29D60B7DE99E00b715309Fd1CbfF3C29F232dd8E";
+  const LP = "0x7EFaEf62fDdCCa950418312c6C91Aef321375A00";
+  const CHAINLINK_TOKEN0_ORACLE = "0xB97Ad0E74fa7d920791E90258A6E2085088b4320";
+  const CHAINLINK_TOKEN1_ORACLE = "0xcBb98864Ef56E9042e7d2efef76141f15731B82f";
 
   await withNetworkFile(async () => {
-    console.log(`>> deploying an OffChainOracle`);
-    const OffChainOracle = (await ethers.getContractFactory("OffChainOracle", deployer)) as OffChainOracle__factory;
-    const offchainOracle = (await upgrades.deployProxy(OffChainOracle)) as OffChainOracle;
+    const pair = LatteSwapPair__factory.connect(LP, deployer);
+    console.log(`deploying an LP Chainlink Aggregator ${await pair.token0()} - ${await pair.token1()} (${LP})`);
+    const LPChainlinkAggregator = (await ethers.getContractFactory(
+      "LPChainlinkAggregator",
+      deployer
+    )) as LPChainlinkAggregator__factory;
+    const lpChainlinkAggregator = (await upgrades.deployProxy(LPChainlinkAggregator, [
+      LP,
+      CHAINLINK_TOKEN0_ORACLE,
+      CHAINLINK_TOKEN1_ORACLE,
+    ])) as LPChainlinkAggregator;
 
-    await offchainOracle.deployed();
-    console.log(`>> Deployed at ${offchainOracle.address}`);
-    console.log("✅ Done deploying OffChainOracle");
-
-    console.log(`>> grant feeder role to ${FEEDER}`);
-    const tx = await offchainOracle.grantRole(await offchainOracle.FEEDER_ROLE(), FEEDER, {
-      gasPrice: ethers.utils.parseUnits("20", "gwei"),
-    });
-    console.log(`✅ Done in tx ${tx.hash}`);
+    await lpChainlinkAggregator.deployed();
+    console.log(`>> Deployed at ${lpChainlinkAggregator.address}`);
+    console.log("✅ Done deploying an LP Chainlink Aggregator");
   });
 };
 
 export default func;
-func.tags = ["DeployOffChainOracle"];
+func.tags = ["DeployLPChainlinkAggregator"];
